[Cases] Toggle columns in the cases table (#169543)
Connected to #167617

This PR is being merged into a feature branch

## Summary

This PR allows users to configure which columns are displayed in the
cases table and in which order.

The configuration is persisted in local storage and is
application-specific.

Custom fields can also be selected.


https://github.com/elastic/kibana/assets/1533137/da386a6c-b2d9-4a86-948f-79903163ff3c

## Coming later in the Feature Branch

The scope of this PR is limited to the Summary above. There are still
some things under consideration and some coming later. These are:
- The display of the toggle in the table. Right now it looks clickable
when it isn't. I will discuss it with @mdefazio and pick a different
design.
- Scrolling inside the popover
- Search inside the popover
- `Show all` and `Hide all` functionality
- In the final PR I will play with the column widths and try to find the
best size for each attribute.

---------

Co-authored-by: kibanamachine <42973632+kibanamachine@users.noreply.github.com>