[Security Solution] [Sourcerer] Adds missing alerts index filtration from sourcerer that was removed during refactoring (#170484)
## Summary

Fixes: https://github.com/elastic/security-team/issues/7859

Thanks to @XavierM and @kqualters-elastic for spending time debugging
this with me.

When we added the [server-side
fetching](https://github.com/elastic/kibana/pull/163448) of index
patterns we missed the functionality that was filtering out the alerts
index from the default sourcerer scope. Without this filtering, the
logic for rendering the matched indices was not correctly refreshing
when adding data the first time.

The result was the alerts index would be set as a part of the filtered
index patterns once a source event data index was present (`logs-*`,
`auditbeat-*` etc..) and the redux store action to set the latest
matched indices was not called because the `useDataView` hook would
incorrectly believe the sourcerer data view no longer needed to be
initialized.




Steps to reproduce:

1. Start local ES and kibana
2. Navigate to Security Solution -> Overview page
3. Welcome / landing page should be visible
4. Start auditbeat (or generate any event data that would be part of the
security solution default data view index patterns)
5. Navigate to Discover, wait for data to load
6. Navigate back to Security Solution -> Overview page
7. Data should be visible, Data View (sourcerer) in the header should
display the correct index pattern for which data exists. Alerts index
should not be included.