Update chromedriver to v116.0.0
Chromedriver downloads moved to a new url scheme around the v114
release.  v116 is the latest version we can upgrade to due to the
package introducing version requirements on Node.js.  On CI, this will
still download the latest version of the chromederiver binary - we
are setting the environment variable DETECT_CHROMEDRIVER_VERSION=true.