[Cases] Enhance Multi Select Component Interface for Custom Fields (#171102)
Meta issue: https://github.com/elastic/kibana/issues/167651

## Summary
We need to make a small update to our multi select component. Right now,
it only allows options with labels, but for custom fields, we might have
situations where different fields have the same label. So, we're going
to adjust it to work with options that have both {key, label}.

For our system filters (status, category, severity, ...) we'll keep
using the label as the key. This is important because it helps us
maintain our current way of handling the global state of our filters
without any issues

## QA
An exhaustive QA is not required for this PR, as it's only going into a
feature branch. The filters should continue working