[Security Solution] [Analyze Event] Make analyze event nodes have stable ids (#171372)
## Summary

During a quick glance to see if the unified fields table/list usage in
timeline were introducing any performance issues, I noticed 2 small
issues that I wanted to fix separately. The first is the ids generated
on the front end and used for different aria-* attributes are not stable
like they should be, and instead change every render. Now they are
generated only once when the node appears on screen, more in line with
how MDN says they should be used.
Before:

![resolver_node_ids_bad](https://github.com/elastic/kibana/assets/56408403/287363cc-c5a7-46e1-8f6b-9dceea829005)
After:

![resolver_node_ids_fixed](https://github.com/elastic/kibana/assets/56408403/671f059b-4fcd-4a99-8df4-a77b86f35e61)

Another component that was needlessly re-rendering on every single
security app redux store change was the alert_context_menu, due to the
"selectors" used to get the global/timeline queries for refetching data
in the different context menu actions being not really selectors at all,
and would instead return a new object every single time. This removes
the react-redux connect method for hooks, and makes the not really a
selector function not change constantly. A lot of our reselect usage
should be overhauled and verified for correctness at some point, but
this is a small improvement. In the gifs above, each mouse movement
fires an event that changes a value in the analyzer part of the store
that tracks mouse position, and the alert context menu component,
including in the underlying alerts table cells if a user is on that
page, would re-render on every single 'onmousemove' event, after this pr
it no longer does.