[Dataset quality] Adding datastreams stats end point (#170838)
Related to https://github.com/elastic/kibana/issues/169759.

## Summary
The Data streams stats endpoint can be queried like so (example with all
parameters):

`GET
/internal/dataset_quality/data_streams/stats?type=logs&sortOrder=asc&datasetQuery=nginx`

### Notes/questions
Our intial idea was to place this endpoint within fleet, after some
discussions we decided to place the endpoint inside our plugin
`dataset_quality`.

- ~~Is `/epm` still the best place to include this new api?~~
    We decided to place the new endpoint in `dataset_quality` plugin.
- ~~Should we include an `uncategorisedOnly` option as part of the Data
Streams Stats endpoint as we do for [Data streams
endpoint](https://github.com/elastic/kibana/blob/c1681c8d49304f324895acbe73323ffacad3fcbc/x-pack/plugins/fleet/server/types/rest_spec/epm.ts#L62)?
We are not planning to have the a filter for this but it could favor
consistency in this area.~~
uncategorizedOnly is not part of the new endpoint although it's easily
extensible to include such parameter.
- ~~We are modifying
`x-pack/plugins/fleet/server/services/epm/data_streams/get.ts` to
include a property for extendedResults, as of now this method only
returns a list of objects that include only the name. We would like to
extend this method to include also information related to the `package`.
Is this the right approach? should we create a new method just for doing
this? Also, this is the method we are using for `GET
/api/fleet/epm/data_streams` since this new property
(`extendedResponse`) is false by default the api is not being directly
affected, but should we include this new param as part of the api
params?~~
We are not working with fleet api anymore, instead we are using an
internal route.
- ~~How does `GetEpmDataStreamsStatsResponse` looks like? is this
reponse type aligned with fleet responses?~~
- ~~We have configured `fleetAuthz: READ_PACKAGE_INFO_AUTHZ` for this
new endpoint, would it be possible to not have this restriction and use
the internal user when querying elastic search?~~

---------

Co-authored-by: kibanamachine <42973632+kibanamachine@users.noreply.github.com>