[ML] Transforms: Create optional data view on server side as part of transform create API call. (#160513)
When creating a transform, users have the option to create a Kibana data
view as part of the creation process. So far we've done that as a
separate call from the client side. This PR adds query param options to
the API endpoint to optionally create the data view server side.
API integration tests have been extended to test the new feature.