[APM] Make data view space aware (#170857)
Closes https://github.com/elastic/kibana/issues/169924
Related SDH: https://github.com/elastic/sdh-apm/issues/1080 (_internal_)

The APM data view is automatically created when opening the APM UI app
via a request to `POST /internal/apm/data_view/static`. This creates a
data view which is made available to all spaces. The problem with this
approach is that users can change index settings (apm index patterns)
per space, which means that there should be a data view per space as
well.

This PR ensures that a data view is created per space and that any links
to the data view uses the spaace specific dataview

---------

Co-authored-by: Caue Marcondes <caue.marcondes@elastic.co>
Co-authored-by: Cauê Marcondes <55978943+cauemarcondes@users.noreply.github.com>