[Security Solution] Fix observable re-rendering component (#171772)
## Summary

There was a "memory-leak" bug on the `ContractComponents` logic (Get
Started page), which was causing a React error.

<img width="814" alt="render_depth_error"
src="https://github.com/elastic/kibana/assets/17747913/c6353e53-9e4a-4103-b61c-1aaf67cda590">

The bug was caused by the `pipe` call in the `getComponent$` function
returning a new observable every call and the `LandingPageComponent` was
not memorizing it.

I changed the `pipe` call by a regular `asObservable`, and created a
reusable `useContractComponents` hook to unify the implementation, which
also memorizes the observable.

I leveraged the changes to the `ContractComponents` implementation to
solve another problem, allowing the stored components to render normally
as a `<Component />`, receiving props if needed, instead of having to
render always as an `{element}`.

---------

Co-authored-by: Kibana Machine <42973632+kibanamachine@users.noreply.github.com>
(cherry picked from commit 49d1b6daaafc39cc284bdbf597418577f06bb403)

# Conflicts:
#	x-pack/plugins/security_solution/public/mocks.ts
#	x-pack/plugins/security_solution/public/plugin_contract.ts
#	x-pack/plugins/security_solution/public/types.ts