[ML] Fixing blocked jobs polling interval (#171878)
There is a bug with the way we check whether or not to start polling for
blocked jobs. We should only start polling there isn't an existing poll
running.
This causes the polling to run as possible and to refresh the full jobs
list on every check.
Now the blocked jobs polling correctly runs at the interval of 2s and
the full jobs list is only updated when there is a change in the list of
blocked jobs.

Partially fixes https://github.com/elastic/kibana/issues/171626

A more in-depth change as specified in the issue can be done in follow
up work. It's good to get this change in for the next release just in
case we run out of time for the larger change.

(cherry picked from commit e7c793e0d74f2c48c788e7daf37dfc3b7cd68c9f)