[CSV/Reporting] Refinements for logging hit totals (#171811)
## Summary

This PR fixes a **logging** issue noticed when analyzing CSV Export
logs. There is an "info" log giving the total number of hits, which
includes the `total.relation` value in the message. This log message
incorrectly printed the relationship as `eq` even when the relation was
not given in the Elasticsearch response.

The correction to this log message will help us make sure our search
requests are configured to track the total hits of the query.

```
# Before
Total hits eq 12345.

# After
Received total hits: 12345. Accuracy: eq.
Received total hits: 12345. Accuracy: unknown.
```

### Other changes
* A new private method for logging the result metadata
* Syntax cleanup
* More tests

### Checklist

Delete any items that are not applicable to this PR.

- [x] [Unit or functional
tests](https://www.elastic.co/guide/en/kibana/master/development-tests.html)
were updated or added to match the most common scenarios