[Response Ops][Alerting] Remove echoed field value from bulk error responses when indexing alerts (#172020)
## Summary

When alerts are bulk indexed in the rule registry and the alerts client,
indexing errors may be returned where the entire field value that failed
to be indexed is echoed in the reason. This can cause unnecessarily
verbose logging so we want to sanitize the field value.

(cherry picked from commit d9ebfd9af1365bba54d5e1ac92f5e53f5fbebea8)