[Docs][Maps] Include details about the headers requested and served by EMS (#171659)
Fixes #129751

## Summary

Extends the EMS documentation to detail request and response headers the
browser sends to EMS resources and includes also a minimal `curl`
command to request the response headers for same resource as well.

I tried to edit this in a way it does not take the whole page but happy
to hear feedback or ideas on how to make this easier to digest.

https://github.com/elastic/kibana/assets/188264/27e83a5f-4d01-47a8-af2c-3739576bf56e

Also, I am not sure if this is something worth adding to our release
notes :thinking:

(cherry picked from commit 0a7299b0ac2aadfe47aa374002636341a032a6ed)

# Conflicts:
#	docs/maps/connect-to-ems.asciidoc