[ES|QL] Make the dashboard SO lighter (#172130)
## Summary

This PR makes the dashboard SO lighter by removing the initialContext
from the state. It is not needed after the navigation from Discover to
Dashboard and can create SO size problems as it contains a lot of
information for the transition.

(cherry picked from commit cb3fd2132f47d43e4d46b8f5b085195cb8dda806)

# Conflicts:
#	x-pack/plugins/lens/public/datasources/text_based/text_based_languages.test.ts