Fleshing out more of the implementation of finding criticality records
Since the AssetCriticalityDataClient already has the context of the
namespace and the ES client, I'm thinking that this function might be
most appropriate there? Not sure yet, I'm still working that out

In summary, this simplifies the interface of the data-fetching function,
while also starting to implement the actual query generation and data
fetching.