[ObsUX] Move formulas and dashboard config to inventory models (#171872)
## Summary

This PR moves formulas and charts developed for the Hosts View and Asset
Details into the Inventory Models existing structure.

### `metrics_data_access` plugin

This is where the inventory models are found. All charts and formulas
have been moved from `infra` plugin into this plugin Most of the changes
are just about that. I've refactored a few things to help with the
reusability of these charts/dashboards

### `infra` plugin

Changes made here were a consequence of the refactors in the
charts/dashboards config.

### `@kbn/lens-embaddable-utils` package

Just added new types

---------

Co-authored-by: kibanamachine <42973632+kibanamachine@users.noreply.github.com>