[Lens] Reduce call stack for formula (#171744)
## Summary

Reduce a bit the call stack with associated perf speedup in a scenario
with many fields.

*Note*: initially I've optimize somehow the memoization of
`getAvailableOperationsByMetadata` but it was too brittle as changing
something in the dataview leads to wrong results. I'm not sure this
could actually happen in real life ™️ but a test hijacking a bit a
dataView found that out.