[Security Solution] Save timeline as new (#171027)
## Summary

This is the second part of the work for fixing timeline saving issues as
described in https://github.com/elastic/kibana/issues/165053. The first
part, replacing auto-save with a dedicated save button was done in this
PR: https://github.com/elastic/kibana/pull/169239.

In this PR we're adding the `Save as new` option to the timeline save
modal. This allows users to create a copy of the timeline without
changing the current timeline. This is especially helpful in a
multi-user environment where there are higher chances for users causing
conflicting changes.

This is how the feature works:


https://github.com/elastic/kibana/assets/68591/dcec9938-fb28-4c3e-b916-8589c6e28441

Notice how the changes to the date range are only applied to the second
timeline. The initial timeline remains unchanged.

### Changes

- Added a new route `/api/timeline/_copy` that expects a timeline id and
a timeline object. It will take the timeline object from the body and
create a new timeline from it. Then it will fetch all related saved
objects (notes, pinned events) for the timeline id in the body and
create copies of those. These will be associated to the copied timeline.
- Creating a copy of the associated saved search of a timeline is not
done on the server because there is currently no server-side API to do
that (there is an issue in the backlog for adding that though:
https://github.com/elastic/kibana/issues/170580).
- This means that the copy is happening in the frontend before making
the copy request. While this feels inconsistent, it's the only way we
can do this at the moment.
- The `Save as new` switch is only shown for timelines that have been
saved before.
- A lot of work was done to make the change-tracking of the embedded
saved search work correctly. This required to split up setting and
initializing the saved search. On top, we're now storing the saved
search object in the local timeline store in order to pass it to the
copy request.
- Various stability and type fixes in the timeline epic
- There's no API documentation for the copy route because OpenAPI
doesn't allow to specify private endpoints. We really don't want users
to use this endpoint yet until the saved search copy can be performed on
the backend.

### Checklist

Delete any items that are not applicable to this PR.

- [x] [Unit or functional
tests](https://www.elastic.co/guide/en/kibana/master/development-tests.html)
were updated or added to match the most common scenarios